move page scrolling functions to helper service
